jobvision_ai/backend/requirements.txt
fastapi
uvicorn
motor
PyJWT
passlib[bcrypt]
python-multipart
email-validator
aiosmtplib
python-dotenv
python-jose
supabase
